requirements.txt
torch>=1.12
pandas
point-e
tqdm
pyyaml